solve first layer of cube
Modified the program to include one solver script, which we might remove
later in favor of individual method calls within run_main.py. Code needs
much more extensive commenting and refactoring, as much of it is not DRY
and relies on many manual keystrokes to run. Goal later will be to get
each script to less than 80 lines of actual code.